Add experimental 3rd cloud algorithm. - This 3rd algorithm is simply a combination of the other two. - It takes the contrast improvements from the first and the four way split thresholding from the other.
This is the best yet, but still faces a slight increase of non cloud objects problem that the first non split algorithm encountered. I've been tweaking it all day and this is what I have so far. I suspect the next step is no longer in improving what clouds get contrasted, but rather finding a way to ignore things that aren't clouds.

The cloud constrating can be even further increased by swapping the isodata thresholding for Li thresholding, but unsurprisingly this also increases the non clouds that get contrasted as well.