Rewrite Mask.py to increase run speed.
- Mask is a speed bottleneck for transforming and cloud contrasting.
- This rewrite firstly makes generate methods return the mask as a binary
image rather than as a list of pixels. This allows apply_mask to simply
multiply the mask with the image to be masked.
- This then allows generate to load a saved mask and then simply convert
binary 1-0 (rather than 255-0) and use that.
- This saves time, since generating a mask from scratch requires looping
through every pixel in 6 512x512 images (rather slow). The new method loads
the mask and immediately returns it.
-I have added a boolean value that forces generate to use the old loop method
everytime it generates, ignoring a saved Mask.py, rather than just the first
time it's run.