Add band removal to Image Classifier.
If the exposure of an image is 0.3s then a mysterious darkened band typically appears.
This is difficult for the network to deal with, and even with training designed to account for this the network
still has trouble accounting for it. This method simply finds the band and brightens it to match the rest of the image.